// src/app/shared/interfaces/defaultdata.interface.ts
// Generated by https://quicktype.io

// Generated by https://quicktype.io
export interface APIResponse<T> {
  success: boolean;
  statusCode: number;
  message: string;
  data: T[];
  token?: string;
}

export enum ReservationStatus {
  Pending,
  Cancelled,
  Rejected,
  Approved,
  Completed
}

export interface Reservation {
  reservationId:    string;
  customer:         Customer;
  table:            Table;
  reservationStart: string;
  reservationEnd:   string;
  numberOfPeople:   number;
  status:           string;
  reservationCode:  string;
  preferences?:      string;
}

export interface UpdReservation {
  tableId:          string;
  reservationStart: string;
  reservationEnd:   string;
  numberOfPeople:   string;
  preferences:      string;
}

export interface CrtReservation {
  tableId:          string;
  reservationStart: string;
  reservationEnd:   string;
  numberOfPeople:   string;
  preferences:      string;
  customer:         CrtCustomer;
}

// Staff
export interface Staff {
  staffId:   string;
  firstName: string;
  lastName:  string;
  email:     string;
}

export interface CrtStaff {
  firstName: string;
  lastName:  string;
  email:     string;
  password:  string;
}

// Customer
export interface Customer {
  customerId:  string;
  fullName:    string;
  email:       string;
  phoneNumber: string;
}

export interface CrtCustomer {
  fullName:    string;
  email:       string;
  phoneNumber: string;
}

// Table
export interface Table {
  tableId:  string;
  name:     string;
  capacity: number;
}

export interface CrtTale {
  name:     string;
  capacity: number;
}